Refactor pt 6.

+++ src/games/zertz/logic.rs
@@ -609,8 +609,8 @@ pub fn apply_capture(
 
     // Compute direction from start and dest
     // dest = start + 2 * direction_offset, so direction_offset = (dest - start) / 2
-    let dy = ((dest_y as i32 - start_y as i32) / 2) as i32;
-    let dx = ((dest_x as i32 - start_x as i32) / 2) as i32;
+    let dy = ((dest_y as i32 - start_y as i32) / 2);
+    let dx = ((dest_x as i32 - start_x as i32) / 2);
 
     // Calculate capture and landing positions
     let cap_y = (start_y as i32 + dy) as usize;
